+++ zarrs/src/array_subset/iterators.rs
@@ -102,7 +102,8 @@ mod tests {
         let indices = subset.contiguous_indices(&[2, 2]).unwrap();
         let mut iter = indices.into_iter();
         assert_eq!(iter.size_hint(), (1, Some(1)));
-        assert_eq!(iter.next(), Some((vec![0, 0], 4)));
+        assert_eq!(iter.contiguous_elements(), 4);
+        assert_eq!(iter.next(), Some(vec![0, 0]));
         assert_eq!(iter.next(), None);
     }
 
@@ -115,8 +116,9 @@ mod tests {
         assert_eq!(indices.contiguous_elements_usize(), 2);
         let mut iter = indices.iter();
         assert_eq!(iter.size_hint(), (2, Some(2)));
-        assert_eq!(iter.next_back(), Some((vec![2, 1], 2)));
-        assert_eq!(iter.next(), Some((vec![1, 1], 2)));
+        assert_eq!(iter.contiguous_elements(), 2);
+        assert_eq!(iter.next_back(), Some(vec![2, 1]));
+        assert_eq!(iter.next(), Some(vec![1, 1]));
         assert_eq!(iter.next(), None);
     }
 
@@ -126,7 +128,8 @@ mod tests {
         let indices = subset.contiguous_indices(&[3, 1, 2, 2]).unwrap();
         let mut iter = indices.into_iter();
         assert_eq!(iter.size_hint(), (1, Some(1)));
-        assert_eq!(iter.next(), Some((vec![1, 0, 0, 0], 8)));
+        assert_eq!(iter.contiguous_elements(), 8);
+        assert_eq!(iter.next(), Some(vec![1, 0, 0, 0]));
         assert_eq!(iter.next(), None);
     }
 
@@ -143,8 +146,9 @@ mod tests {
         //  8  9 10 11
         // 12 13 14 15
         assert_eq!(iter.size_hint(), (2, Some(2)));
-        assert_eq!(iter.next_back(), Some((9, 2)));
-        assert_eq!(iter.next(), Some((5, 2)));
+        assert_eq!(iter.contiguous_elements(), 2);
+        assert_eq!(iter.next_back(), Some(9));
+        assert_eq!(iter.next(), Some(5));
         assert_eq!(iter.next(), None);
     }